Don't load product_variables.bzl in cc_toolchain_features.bzl
product_variables.bzl should not be loaded directly, but instead
read from the transitionable //build/bazel/product_config:product_vars
target.

Some of the information in product_variables.bzl is constant, and
doesn't need to be read from that target. In those cases, load the
information from other files so that the product_variables.bzl file
can be made private.

Bug: 269577299
Test: m nothing
Change-Id: I02e38c9f2c79143b014936abab58a721759d70ab