exclude ubsan rt globally by default
This change removes the requirement to have the `ubsan_enabled` feature enabled for the `exlcude-libs` flag to be added to the link action.

While this isn't the ideal situation, we don't have a good way in our current macro-based rules to inform a cc_library_shared that it must exclude the ubsan runtime if its deps use ubsan.

Instead, this change opts to just always exclude the ubsan runtime since there shouldn't be any negative consequences for libraries that don't use ubsan.

Bug: 280348200
Test: b build //frameworks/av/apex:com.android.media.swcodec &&
  diff \
  <(readelf -s --demangle --wide \ bazel-out/aosp_arm64-eng-opt-ST-54653e5b7b44/bin/system/memory/libdmabufheap/liblibdmabufheap_unstripped.so \ | awk '{print $8}' | sort -u) \
  <(readelf -s --demangle --wide out/soong/.intermediates/system/memory/libdmabufheap/libdmabufheap/android_arm64_armv8-a_shared_apex31/unstripped/libdmabufheap.so | awk '{print $8}' | sort -u)
Change-Id: I27b4fc1e00733a5d8d02d20a698719460b533a5d