Symlink to checked-in prebuilt of libncurses.so.5
libncurses.so.5 is an older dependency of the RenderScript Clang, but it
is also likely to continue being removed by newer versions of Linux. As
part of the RenderScript deprecation, this is part of a 2 step process
to remove the legacy RenderScript Clang. For part one, I'm switching our
dependency over to a checked-in prebuilt of the ncurses5 library. Next,
I'll create a final set of RenderScript runtime libraries using this
Clang on a buildbot, and document how we would regenerate them if needed
in the future.

Bug: http://b/301495450
Test: sudo apt remove libncurses5; cd frameworks/rs/driver/runtime && mm
Change-Id: I3b0d4f09e44f5b46d4ae6362af86d22619dcf2c3