Remove std::allocator<const T> extension
Cherry-pick llvm.org/PR96319, which removes the extension. The removal
mostly affects code using a const element type in an STL container,
such as std::vector, std::set, std::queue, or std::list.

The cherry-picked patch makes the extension opt-in using the
_LIBCPP_ENABLE_REMOVED_ALLOCATOR_CONST macro, which is respected in
upstream LLVM 19, but for later LLVM releases (and in current upstream
LLVM trunk), the extension is removed completely with no opt-in macro.

Cherry-pick the patch into the clang-r530567 prebuilt to catch users of
the extension and to prevent new ones. Because this patch only affects
the prebuilt, it does not include the fixes to tests from
llvm.org/PR96319.

Bug: http://b/349681543
Test: treehugger
Flag: EXEMPT, refactor to fix build failure
Change-Id: I136ba4363073fbe21995739fe21d835da4b3e017